Work in progress on the booking interface do not touch pls

+++ database/migrations/2020_07_11_070036_create_atc_roster_members_table.php
@@ -14,7 +14,7 @@ class CreateAtcRosterMembersTable extends Migration
     public function up()
     {
         Schema::create('atc_roster_members', function (Blueprint $table) {
-            $table->integer('id')->unique();
+            $table->bigInteger('id')->unique();
             $table->integer('vatsim_id');
             $table->string('fname');
             $table->string('lname');
